# api/openapi.yaml
openapi: 3.0.0
info:
  title: Library Management API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        stock:
          type: integer
    Member:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: 
          type: string
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    BorrowInput:
      type: object
      required: [member_id, book_id]
      properties:
        member_id:
          type: integer
        book_id:
          type: integer
    TokenResponse:
      type: object
      properties:
        token:
          type: string

paths:
  /auth/register:
    post:
      summary: Register a new member
      operationId: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        200:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'

  /auth/login:
    post:
      summary: Login
      operationId: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

  /books:
    get:
      summary: List books
      security:
        - BearerAuth: []
      operationId: GetBooks
      responses:
        200:
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      summary: Create book
      security:
        - BearerAuth: []
      operationId: CreateBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        200:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'

  /borrow:
    post:
      summary: Borrow a book
      security:
        - BearerAuth: []
      operationId: BorrowBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowInput'
      responses:
        200:
          description: Success